[REFAC] 상단바 뒤로가기 버튼 수정
