+++ StockMate/StockMate/app/core/components/CategoryButton.swift
@@ -1,0 +1,32 @@
+//
+//  CategoryButton.swift
+//  StockMate
+//
+//  Created by Admin on 11/8/25.
+//
+
+import SwiftUI
+
+struct CategoryButton: View {
+    let title: String
+    let isSelected: Bool
+    let action: () -> Void
+    
+    var body: some View {
+        Button(action: action) {
+            Text(title)
+                .font(.system(size: 13, weight: .regular))
+                .foregroundColor(isSelected ? .Primary : .black)
+                .padding(.vertical, 8)
+                .padding(.horizontal, 12)
+                .background(
+                    RoundedRectangle(cornerRadius: 16)
+                        .fill(Color.Light)
+                )
+                .overlay(
+                    RoundedRectangle(cornerRadius: 16)
+                        .stroke(isSelected ? Color.Primary : Color.GrayStroke, lineWidth: 1)
+                )
+        }
+    }
+}
