+++ StockMate/StockMate/app/feature/dashboard/ui/ReleaseDetailView.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 struct ReleaseDetailView: View {
+    @Environment(\.dismiss) private var dismiss
     let history: HistoryItem
 
     var body: some View {
@@ -35,6 +36,20 @@ struct ReleaseDetailView: View {
         .background(Color.Light)
         .navigationTitle("출고 상세")
         .navigationBarTitleDisplayMode(.inline)
+        .navigationBarBackButtonHidden(true)
+        .toolbar {
+            ToolbarItem(placement: .navigationBarLeading) {
+                Button {
+                    dismiss()
+                } label: {
+                    HStack(spacing: 4) {
+                        Image(systemName: "arrow.left")
+                            .font(.system(size: 15, weight: .medium))
+                    }
+                    .foregroundColor(.black)
+                }
+            }
+        }
     }
 }
