+++ StockMate/StockMate/app/feature/notification/ui/NotificationListView.swift
@@ -7,6 +7,7 @@
 import SwiftUI
 
 struct NotificationListView: View {
+    @Environment(\.dismiss) private var dismiss
     @StateObject private var notificationViewModel = NotificationViewModel()
     @State private var selectedOrderId: Int? = nil
 
@@ -29,7 +30,19 @@ struct NotificationListView: View {
         .background(Color.Light)
         .navigationTitle("알림")
         .navigationBarTitleDisplayMode(.inline)
+        .navigationBarBackButtonHidden(true)
         .toolbar {
+            ToolbarItem(placement: .navigationBarLeading) {
+                Button {
+                    dismiss()
+                } label: {
+                    HStack(spacing: 4) {
+                        Image(systemName: "arrow.left")
+                            .font(.system(size: 15, weight: .medium))
+                    }
+                    .foregroundColor(.black)
+                }
+            }
             ToolbarItem(placement: .navigationBarTrailing) {
                 Button("전체 읽음") {
                     Task { await notificationViewModel.markAllAsRead() }
